Move reftest code to test_browser.py. NFC
Hopefully the last in my long chain of recent reftest changes.
